[证据线] fix(backend): 蓝图 upsert 三态计数修正——Added/Updated/Kept 枚举替代 bool,修复重复解析误报 nodes_updated,docstring 同步

+++ backend/src/models/kg_blueprint.rs
@@ -98,8 +98,8 @@ pub struct BlueprintDocumentRead {
     pub title: String,
     pub source_kind: String,
     pub parse_mode: String,
-    pub node_count: i64,
-    pub edge_count: i64,
+    pub node_count: i32,
+    pub edge_count: i32,
     pub message: String,
     pub created_at: DateTime<Utc>,
 }
